% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_df.R
\name{bind_df}
\alias{bind_df}
\title{Binding the Rows of Many Dataframes Together}
\usage{
bind_df(input)
}
\arguments{
\item{input}{Pattern (regex) of the dataframes that you want to row bind.}
}
\description{
This function will bind the rows of all dataframes together that match a certain pattern
in the global environment.
}
\examples{

 dd1 <- data.frame(a = 1, b = 2)
 dd2 <- data.frame(a = 3, b = 4)
 dd3 <- data.frame(a = 5, b = 6)

bind_df("dd")
}
\keyword{bind_rows}
