% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{ct}
\alias{ct}
\title{A Count Function with Weights}
\usage{
ct(df, var, wt, cum = FALSE, show_na = TRUE)
}
\arguments{
\item{df}{Name of the Dataset}

\item{var}{Variable to Count}

\item{wt}{Weighting Variable}

\item{cum}{Will add a cumulative total if set to TRUE}

\item{show_na}{Will remove the NAs before the calculation if set to TRUE, defaults to FALSE}
}
\description{
This function gives you a simple count with percentages that is weighted
}
\examples{
ct()
}
\keyword{Count}
