% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_bar.R
\name{error_bar}
\alias{error_bar}
\title{Error Bars for mean_ci() Output}
\usage{
error_bar(wd = 0.2)
}
\arguments{
\item{wd}{Width of the error bar caps. Default is `0.2`.}
}
\value{
A ggplot2 layer that can be added to a plot with `+`.
}
\description{
A shortcut for [ggplot2::geom_errorbar()] that maps the `lower` and `upper`
columns produced by [mean_ci()], with dodging that matches a standard
dodged bar chart.
}
\details{
The data must contain columns named `lower` and `upper`, as returned by
[mean_ci()]. The layer uses `position_dodge(0.9)` so the bars line up with
`geom_col(position = "dodge")` defaults.
}
\examples{
library(dplyr)
library(ggplot2)

mtcars \%>\%
  group_by(cyl) \%>\%
  mean_ci(mpg) \%>\%
  ggplot(aes(x = factor(cyl), y = mean)) +
  geom_col() +
  error_bar()

}
