% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_zero.R
\name{na_zero}
\alias{na_zero}
\title{Converts NAs in a dataset to zeroes}
\usage{
na_zero(df, var)
}
\arguments{
\item{df}{Name of the Dataset}

\item{var}{Variable to find convert NAs to Zero}
}
\description{
This function will convert all the NA values in a vector to 0
}
\examples{
na_zero()
}
\keyword{Mean}
