% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbar.R
\name{xbar}
\alias{xbar}
\title{A Crosstab Stacked Bar Graph}
\usage{
xbar(df, var1, var2, count = TRUE)
}
\arguments{
\item{df}{Name of the Dataset}

\item{var1}{This will be the variable that you do group_by with}

\item{var2}{This is the variable that will be counted}

\item{count}{Will add the total count to each square if added}
}
\description{
This function creates a crosstabulation heatmap
}
\examples{

mtcars \%>\% xbar(cyl, am)


}
\keyword{Crosstab,}
\keyword{bar}
\keyword{graph}
\keyword{stacked}
