% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_df.R
\name{bind_df}
\alias{bind_df}
\title{Binding the Rows of Many Dataframes Together}
\usage{
bind_df(input)
}
\arguments{
\item{input}{Pattern of the dataframes that you want to row bind}
}
\description{
This function will bind the rows of all dataframes together that match a certain pattern
}
\examples{
bind_df()
}
\keyword{bind_rows}
