% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr}
\alias{corr}
\title{Correlation with p-values}
\usage{
corr(df, var1, var2)
}
\arguments{
\item{df}{Name of the Dataset}

\item{var1}{First variable}

\item{var2}{Second variable}
}
\description{
This function gives you a mean with 95 percent CIs
}
\examples{
mean_ci()
}
\keyword{correlation,}
\keyword{p-value}
