% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{ct}
\alias{ct}
\title{Count and tabulate with percentages (weighted or unweighted)
Respects existing dplyr groups and computes percentages within each group.}
\usage{
ct(df, var, wt, show_na = TRUE, cum = FALSE, sort = FALSE)
}
\description{
Count and tabulate with percentages (weighted or unweighted)
Respects existing dplyr groups and computes percentages within each group.
}
