% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{ct}
\alias{ct}
\title{A Count Function with Weights}
\usage{
ct(df, var, wt, cum = FALSE, show_na = TRUE)
}
\arguments{
\item{df}{Name of the Dataset}

\item{var}{Variable to Count}

\item{wt}{Weighting Variable}

\item{cum}{Will add a cumulative total if set to TRUE}

\item{show_na}{Will remove the NAs before the calculation if set to TRUE, defaults to FALSE}
}
\description{
This function gives you a simple count with percentages that is weighted
}
\examples{


cces <- read_csv("https://raw.githubusercontent.com/ryanburge/blocks/master/cces.csv")
cces \%>\% 
   ct(race)
   
# With a weight 
cces \%>\% 
     ct(race, wt = commonweight_vv)            
     
cces \%>\% 
   mutate(race2 = frcode(race == 1 ~ "White",
                     race == 2 ~ "Black", 
                     race == 3 ~ "Hispanic",
                      race == 4 ~ "Asian")) \%>\% 
  ct(race2, show_na = FALSE, commonweight_vv)   
     
# Cumulative Counts       

cces \%>\% 
     ct(race, cum = TRUE)

}
\keyword{Count}
