% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frcode.R
\name{frcode}
\alias{frcode}
\title{Recodes Variables and Keeps Factor Levels}
\usage{
frcode(...)
}
\arguments{
\item{df}{Name of the Dataset}
}
\description{
This a Modification of case_when that keeps things in the order they were recoded
}
\examples{

cces <- read_csv("https://raw.githubusercontent.com/ryanburge/cces/master/CCES\%20for\%20Methods/small_cces.csv")

cces \%>\% 
    mutate(pid_new = frcode(pid7 == 1 ~ "Strong Democrat", 
                            pid7 == 2 ~ "Not Strong Democrat", 
                            pid7 == 3 ~ "Lean Democrat", 
                            pid7 == 4 ~ "Independent", 
                            pid7 == 5 ~ "Lean Republican", 
                            pid7 == 6 ~ "Not Strong Republican", 
                            pid7 == 7 ~ "Strong Republican", 
                            TRUE ~ "REMOVE")) \%>\% 
           ct(pid_new) \%>\% 
           filter(pid_new != "REMOVE") \%>\% 
           ggplot(., aes(x = pid_new, y = pct)) +
           geom_col()
  

}
\keyword{Mean}
