% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frcode.R
\name{frcode}
\alias{frcode}
\title{Recodes Variables and Keeps Factor Levels}
\usage{
frcode(...)
}
\arguments{
\item{df}{Name of the Dataset}
}
\description{
This a Modification of case_when that keeps things in the order they were recoded
}
\examples{
frcode()
}
\keyword{Mean}
