% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ci.R
\name{mean_ci}
\alias{mean_ci}
\title{A Mean with 95 percent CIs}
\usage{
mean_ci(df, var, wt, ci = 0.95)
}
\arguments{
\item{df}{Name of the Dataset}

\item{var}{Variable to find the mean of}

\item{wt}{Weight to be applied}

\item{ci}{Confidence Interval, expressed as a decimal. i.e. .84. Defaults to .95}
}
\description{
This function gives you a mean with 95 percent CIs
}
\examples{
mean_ci()
}
\keyword{Mean}
