% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_med.R
\name{mean_med}
\alias{mean_med}
\title{Gives You the Mean and Median}
\usage{
mean_med(df, var)
}
\arguments{
\item{df}{Name of the Dataset}

\item{var}{Variable to find the mean and median of}
}
\description{
This function gives you the mean and the median of a dataset
}
\examples{
mean_med()
}
\keyword{Mean}
