% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xheat.R
\name{xheat}
\alias{xheat}
\title{A Crosstab Heatmap}
\usage{
xheat(df, var1, var2, count = FALSE)
}
\arguments{
\item{df}{Name of the Dataset}

\item{var1}{This will be the variable that you do group_by with}

\item{var2}{This is the variable that will be counted}

\item{count}{Will add the total count to each square, it's off by default.}
}
\description{
This function creates a crosstabulation heatmap
}
\examples{

mtcars \%>\% xheat(cyl, am)



}
\keyword{Crosstab,}
\keyword{heatmap}
